+++ simple-stack-example-basic-java-fragment/src/main/java/com/zhuinden/navigationexamplefrag/MainActivity.java
@@ -100,7 +100,7 @@ public class MainActivity
 
     @Override
     public void handleStateChange(@NonNull StateChange stateChange, @NonNull Callback completionCallback) {
-        if(stateChange.topNewState().equals(stateChange.topPreviousState())) {
+        if(stateChange.isTopNewStateEqualToPrevious()) {
             completionCallback.stateChangeComplete();
             return;
         }