+++ simple-stack-example-multistack/src/main/java/com/zhuinden/simplestackdemomultistack/application/MainActivity.java
@@ -177,7 +177,7 @@ public class MainActivity
 
     @Override
     public void handleStateChange(@NonNull StateChange stateChange, @NonNull Callback completionCallback) {
-        if(stateChange.topNewState().equals(stateChange.topPreviousState())) {
+        if(stateChange.isTopNewStateEqualToPrevious()) {
             // no-op
             completionCallback.stateChangeComplete();
             return;