+++ simple-stack-example-mvp-view/src/main/java/com/zhuinden/simplestackdemoexamplemvp/application/MainActivity.kt
@@ -81,7 +81,7 @@ class MainActivity : AppCompatActivity(), StateChanger {
     }
 
     override fun handleStateChange(stateChange: StateChange, completionCallback: StateChanger.Callback) {
-        if (stateChange.topNewState<Any>() == stateChange.topPreviousState<Any>()) {
+        if (stateChange.isTopNewStateEqualToPrevious) {
             completionCallback.stateChangeComplete()
             return
         }